+++ yampt.qt/source/settings_store.hpp
@@ -153,6 +153,9 @@ public:
 	int display_codepage() const;
 	void set_display_codepage(int value);
 
+	bool editing_enabled() const;
+	void set_editing_enabled(bool value);
+
 	bool clean_evil_gmst_enabled() const;
 	void set_clean_evil_gmst_enabled(bool value);